backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.20.0
Pillow==10.4.0
numpy==2.3.2
scikit-learn==1.7.1
nltk==3.8.1
pandas==2.3.2
gunicorn==21.2.0
psutil==7.0.0
python-dotenv==1.0.0
